Multi-threading implementation
Separate BLE and motor control into separate threads.

This seems to work more reliably. I don't think we're quite where we want in terms of synchronization, need to work out a way to make sure both gloves are on the same page about the pattern in a way that is more by design than by chance.